requirements.txt
capstone
unicorn
llvmlite
pyelftools
pytest
